I have no idea what I'm doing with typescript :)
